Move DotDict to snippets
