+++ pyiron_workflow/io.py
@@ -21,7 +21,7 @@ from pyiron_workflow.channels import (
 from pyiron_workflow.has_channel import HasChannel
 from pyiron_workflow.has_to_dict import HasToDict
 from pyiron_workflow.snippets.logger import logger
-from pyiron_workflow.util import DotDict
+from pyiron_workflow.snippets.dotdict import DotDict
 
 
 class IO(HasToDict, ABC):
