+++ tests/unit/test_workflow.py
@@ -5,7 +5,7 @@ import unittest
 
 from pyiron_workflow._tests import ensure_tests_in_python_path
 from pyiron_workflow.channels import NotData
-from pyiron_workflow.util import DotDict
+from pyiron_workflow.snippets.dotdict import DotDict
 from pyiron_workflow.workflow import Workflow
